chore: bump rust 1.91 and edition 2024 & fix all lint issues in compiler_base

+++ compiler_base/3rdparty/rustc_data_structures/src/flock.rs
@@ -38,7 +38,7 @@ cfg_if! {
                     .read(true)
                     .write(true)
                     .create(create)
-                    .mode(libc::S_IRWXU as u32)
+                    .mode(libc::S_IRWXU)
                     .open(p)?;
 
                 let mut operation = if exclusive {
